refactor(motor): isolate registry backend selection and gate motor shell
Agent-Logs-Url: https://github.com/kabot-io/kabot-zephyr/sessions/64d6b8e6-3aba-4d82-8623-caa7b8c51291
