+++ app/prj.conf
@@ -1,4 +1,5 @@
 CONFIG_SHELL=y
+CONFIG_KABOT_MOTOR_SHELL=y
 
 CONFIG_EEPROM=y
 # CONFIG_EEPROM_SHELL=y
